requirements.txt
streamlit==1.7.0
pytesseract==0.3.9
pdf2image==1.16.0
Pillow==9.0.1
wordcloud==1.8.1

